MCP Server/requirements.txt
fastmcp
pandas
langchain
langchain-chroma
langchain-google-genai
python-dotenv
chromadb
langchain-text-splitters
PyPDF2
fastparquet